Number Cross 3: https://www.janestreet.com/puzzles/number-cross-3/
Final solution. Initial bit copied from the sudoku example out of the
SBV docs and then modified to actually be for the correct solution.
Requires z3 or some other SMT solver to run.

Works very quickly, I sort of wish I had an all haskell solution to
compare it with.

Not totally clean (mixes the problem with the checks much more badly
than the sudoku one did) and occasionally does things that look silly,
but an educational experience. Figuring out how to filter and 'sort' the
lists was quite 'satisfying'. oh ho ho!